+++ lib/rubocop/cop/style/indent_hash.rb
@@ -46,17 +46,36 @@ module RuboCop
           return if ignored_node?(hash_node)
 
           first_pair = hash_node.children.first
-          return if first_pair.nil?
+          if first_pair
+            left_brace = hash_node.loc.begin
+            return if first_pair.loc.expression.line == left_brace.line
 
-          left_brace = hash_node.loc.begin
-          return if first_pair.loc.expression.line == left_brace.line
-
-          if separator_style?(first_pair)
-            check_based_on_longest_key(hash_node.children, left_brace,
-                                       left_parenthesis)
-          else
-            check_first_pair(first_pair, left_brace, left_parenthesis, 0)
+            if separator_style?(first_pair)
+              check_based_on_longest_key(hash_node.children, left_brace,
+                                         left_parenthesis)
+            else
+              check_first_pair(first_pair, left_brace, left_parenthesis, 0)
+            end
           end
+
+          check_right_brace(hash_node.loc.end, left_brace, left_parenthesis)
+        end
+
+        def check_right_brace(right_brace, left_brace, left_parenthesis)
+          return if right_brace.source_line[0...right_brace.column] =~ /\S/
+
+          expected_column = base_column(left_brace, left_parenthesis)
+          @column_delta = expected_column - right_brace.column
+          return if @column_delta == 0
+
+          msg = if style == :special_inside_parentheses && left_parenthesis
+                  'Indent the right brace the same as the first position ' \
+                  'after the preceding left parenthesis.'
+                else
+                  'Indent the right brace the same as the start of the line ' \
+                  'where the left brace is.'
+                end
+          add_offense(right_brace, right_brace, msg)
         end
 
         def separator_style?(first_pair)
@@ -75,8 +94,9 @@ module RuboCop
 
         def check_first_pair(first_pair, left_brace, left_parenthesis, offset)
           column = first_pair.loc.expression.column
-          @column_delta =
-            expected_column(left_brace, left_parenthesis, offset) - column
+          expected_column = base_column(left_brace, left_parenthesis) +
+            IndentationWidth::CORRECT_INDENTATION + offset
+          @column_delta = expected_column - column
 
           if @column_delta == 0
             correct_style_detected
@@ -93,16 +113,12 @@ module RuboCop
           end
         end
 
-        # Returns the expected, or correct indentation.
-        def expected_column(left_brace, left_parenthesis, offset)
-          base_column =
-            if left_parenthesis && style == :special_inside_parentheses
-              left_parenthesis.column + 1
-            else
-              left_brace.source_line =~ /\S/
-            end
-
-          base_column + IndentationWidth::CORRECT_INDENTATION + offset
+        def base_column(left_brace, left_parenthesis)
+          if left_parenthesis && style == :special_inside_parentheses
+            left_parenthesis.column + 1
+          else
+            left_brace.source_line =~ /\S/
+          end
         end
 
         # Returns the description of what the correct indentation is based on.